refactor: Ajuste na criação da venda

+++ src/sales/model/saleModel.go
@@ -11,9 +11,9 @@ import (
 type Sale struct {
 	CustomerCareID uuid.UUID                      `gorm:"column:customer_care_id;primarykey" json:"id"`
 	CustomerCare   customerCareModel.CustomerCare `gorm:"foreignKey:CustomerCareID" json:"customer_care"`
-	TotalValue    int       `json:"total_value"`
-	Status        string    `gorm:"type:varchar(45)" json:"status"`
-	PaymentMethod string    `gorm:"type:varchar(45)" json:"payment_method"`
-	CreatedAt     time.Time `json:"created_at"`
-	UpdatedAt     time.Time `json:"updated_at"`
+	TotalValue     float32                        `json:"total_value"`
+	Status         string                         `gorm:"type:varchar(45)" json:"status"`
+	PaymentMethod  string                         `gorm:"type:varchar(45)" json:"payment_method"`
+	CreatedAt      time.Time                      `json:"created_at"`
+	UpdatedAt      time.Time                      `json:"updated_at"`
 }
